# Improved KV cache loading performance for Vulkan, resulting in a 20x …