@0cc4m 2025-02-15
Interesting. There is probably higher overhead on Vulkan because async memory transfer operations are not implemented. I don't think this can be merged in this form, but we should definitely discuss possible solutions (@slaren )