@slaren 2025-02-16
The problem in this case is not that the KV cache is loaded layer by layer, this PR does not change that, and it still requires one call to `ggml_backend_tensor_set` per layer. Without flash attention, the V cache is transposed, which results in `n_embd_v` transfers per layer to restore it, and that is what this change may improve, by building a complete copy of the V cache in memory first and then transferring it at once. However, in cases where the size of the cache being loaded is small compared to the total size of the cache, this will result in a much larger amount of data transferred to the GPU, so I don't think this change would be universally faster, in addition to increasing memory usage.

Supporting flash attention would solve this problem by allowing the use of a non-transposed V cache, then a single `ggml_backend_tensor_set` call per layer would be enough. So I believe it would be preferable to work in that direction instead. Using async memory transfers could also help, but it would also require loading the entire cache into system memory first, increasing memory usage.